Disburesement Cron Job

+++ src/main/kotlin/com/entryventures/models/dto/LoanDto.kt
@@ -7,6 +7,6 @@ import jakarta.validation.constraints.NotBlank
 data class LoanDto(
     @JsonProperty("id") var id: String = "",
     @JsonProperty("status") var status: LoanStatus = LoanStatus.Pending,
-    @JsonProperty("amount") var amount: Int,
+    @JsonProperty("amount") var amount: Float,
     @JsonProperty("client_id") @NotBlank var clientId: String
 )
